app/requirements.txt
affine==2.4.0
asgiref==3.6.0
async-timeout==5.0.1
attrs==22.2.0
boto3==1.28.61
botocore==1.31.61
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Django==4.2
django-colorfield==0.11.0
django-geojson==4.0.0
django-storages==1.14.1
djangorestframework==3.14.0
drf-yasg==1.21.7
earthengine-api==0.1.341
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==21.2.0
hiredis==3.0.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
numpy==1.24.2
packaging==23.2
Pillow==10.0.1
protobuf==4.22.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.2
rasterio==1.3.6
redis==5.2.0
requests==2.28.2
rsa==4.9
s3transfer==0.7.0
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
